[Core/GameObjects] Remove bad casts in relation to GameObject m_unique_users

+++ src/game/GameObject.h
@@ -819,7 +819,7 @@ class GameObject : public WorldObject, public GridObject<GameObject>
         std::list<uint32> m_SkillupList;
 
         Player* m_ritualOwner;                              // used for GAMEOBJECT_TYPE_SUMMONING_RITUAL where GO is not summoned (no owner)
-        std::set<uint32> m_unique_users;
+        std::set<uint64> m_unique_users;
         uint32 m_usetimes;
 
         uint32 m_DBTableGuid;                               // For new or temporary gameobjects is 0 for saved it is lowguid
